bib_lite_agent/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
openai==1.54.3
python-dotenv==1.0.1
